增加合同开启资质审核开关

+++ ehnextgen-api/src/main/java/com/everhomes/customer/EnterpriseCustomerProvider.java
@@ -234,5 +234,5 @@ public interface EnterpriseCustomerProvider {
     CustomerTalent findPotentialCustomerById(Long sourceId);
     List<CustomerAdminRecord> listEnterpriseCustomerAdminRecordsByToken(Long id, String adminToken);
 
-    void updateCustomerAptitudeFlag(Long id, Integer approvalStatus);
+    void updateCustomerAptitudeFlag(Long id, Byte approvalStatus);
 }
